1. npm procedure refactoring 2. set openharmony as remote source 3. check mark of sha256 instead of calculating it

+++ prebuilts_download.sh
@@ -61,6 +61,20 @@ function check_sha256() {
     fi
 }
 
+function check_sha256_by_mark() {
+    success_color='\033[1;42mSuccess\033[0m'
+    check_url=$1 # source URL
+    check_sha256=$(curl -s -k ${check_url}.sha256)
+    echo $1
+    if [ -f "${code_dir}/${unzip_dir}/${check_sha256}.mark" ];then
+        echo -e "${success_color},${check_url} Sha256 markword check OK."
+        sha256_result=0
+    else
+        echo -e "${check_url} Sha256 mismatch or files not downloaded yet."
+        sha256_result=1
+    fi
+}
+
 function hwcloud_download() {
     # when proxy certfication not required : wget -t3 -T10 -O ${bin_dir} -e "https_proxy=http://domain.com:port" ${huaweicloud_url}
     # when proxy certfication required (special char need URL translate, e.g '!' -> '%21'git
@@ -91,6 +105,37 @@ Local sha256: ${local_sha256}"""
     exit 1
 }
 
+function npm_install() {
+    full_code_path=${code_dir}/$1
+    if [ ! -d ${full_code_path} ]; then
+        echo "${full_code_path} not exist, it shouldn't happen, pls check..."
+    else
+        cd ${full_code_path}
+        export PATH=${code_dir}/prebuilts/build-tools/common/nodejs/${node_js_name}/bin:$PATH
+        npm config set registry ${npm_registry}
+        if [ "X${SKIP_SSL}" == "XYES" ];then
+            npm config set strict-ssl false
+        fi
+        # npm cache clean -f
+        npm install
+    fi
+}
+
+function node_modules_copy() {
+    full_code_path=${code_dir}/$1
+    tool_path=$2
+    if [ -d "${full_code_path}" ] & [ ! -z ${tool_path} ]; then
+        if [ -d "${code_dir}/${tool_path}" ]; then
+            echo -e "\n"
+            echo "${code_dir}/${tool_path} already exist, it will be replaced with node-${node_js_ver}"
+            /bin/rm -rf ${code_dir}/${tool_path}
+            echo -e "\n"
+        fi
+        mkdir -p ${code_dir}/${tool_path}
+        /bin/cp -R ${full_code_path}/node_modules ${code_dir}/${tool_path}
+    fi
+}
+
 case $(uname -s) in
     Linux)
         host_platform=linux
@@ -111,37 +156,43 @@ code_dir=$(dirname ${script_path})
 bin_dir=${code_dir}/../OpenHarmony_2.0_canary_prebuilts
 
 # download file list
+# todo: add product related config
 copy_config="""
-prebuilts/sdk/js-loader/build-tools,${tool_repo}/harmonyos/compiler/ace-loader/1.0/ace-loader-1.0.tar.gz
-prebuilts/build-tools/common,${tool_repo}/harmonyos/compiler/restool/2.007/restool-2.007.tar.gz
-prebuilts/cmake,${tool_repo}/harmonyos/compiler/cmake/3.16.5/${host_platform}/cmake-${host_platform}-x86-3.16.5.tar.gz
-prebuilts/build-tools/${host_platform}-x86/bin,${tool_repo}/harmonyos/compiler/gn/1717/${host_platform}/gn-${host_platform}-x86-1717.tar.gz
-prebuilts/build-tools/${host_platform}-x86/bin,${tool_repo}/harmonyos/compiler/ninja/1.10.1/${host_platform}/ninja-${host_platform}-x86-1.10.1.tar.gz
-prebuilts/python,${tool_repo}/harmonyos/compiler/python/3.8.5/${host_platform}/python-${host_platform}-x86-3.8.5.tar.gz
-prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/harmonyos/compiler/clang/10.0.1-480513/${host_platform}/clang-480513-${host_platform}-x86_64.tar.bz2
-prebuilts/ark_tools,${tool_repo}/harmonyos/compiler/llvm_prebuilt_libs/ark_js_prebuilts_20220425.tar.gz
+prebuilts/sdk/js-loader/build-tools,${tool_repo}/openharmony/compiler/ace-loader/1.0/ace-loader-1.0.tar.gz
+prebuilts/build-tools/common,${tool_repo}/openharmony/compiler/restool/2.007/restool-2.007.tar.gz
+prebuilts/cmake,${tool_repo}/openharmony/compiler/cmake/3.16.5/${host_platform}/cmake-${host_platform}-x86-3.16.5.tar.gz
+prebuilts/build-tools/${host_platform}-x86/bin,${tool_repo}/openharmony/compiler/gn/1717/${host_platform}/gn-${host_platform}-x86-1717.tar.gz
+prebuilts/build-tools/${host_platform}-x86/bin,${tool_repo}/openharmony/compiler/ninja/1.10.1/${host_platform}/ninja-${host_platform}-x86-1.10.1.tar.gz
+prebuilts/python,${tool_repo}/openharmony/compiler/python/3.8.5/${host_platform}/python-${host_platform}-x86-3.8.5.tar.gz
+prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/openharmony/compiler/clang/10.0.1-480513/${host_platform}/clang-480513-${host_platform}-x86_64.tar.bz2
+prebuilts/ark_tools,${tool_repo}/openharmony/compiler/llvm_prebuilt_libs/ark_js_prebuilts_20220425.tar.gz
+"""
+
+linux_copy_config="""
+prebuilts/cmake,${tool_repo}/openharmony/compiler/cmake/3.16.5/windows/cmake-windows-x86-3.16.5.tar.gz
+prebuilts/mingw-w64/ohos/linux-x86_64,${tool_repo}/openharmony/compiler/mingw-w64/7.0.0/clang-mingw.tar.gz
+prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi,${tool_repo}/openharmony/compiler/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi/1.0/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi.tar.gz
+prebuilts/gcc/linux-x86/aarch64,${tool_repo}/openharmony/compiler/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi/1.0/gcc-linaro-7.5.0-2019.12-x86_64_aarch64-linux-gnu.tar.xz
+prebuilts/previewer/windows,${tool_repo}/openharmony/develop_tools/previewer/3.1.5.4/previewer-3.1.5.4.win.tar.gz
+prebuilts/clang/ohos/windows-x86_64,${tool_repo}/openharmony/compiler/clang/10.0.1-480513/windows/clang-480513-windows-x86_64.tar.bz2
+prebuilts/clang/ohos/windows-x86_64,${tool_repo}/openharmony/compiler/clang/10.0.1-480513/windows/libcxx-ndk-480513-windows-x86_64.tar.bz2
+prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/openharmony/compiler/clang/10.0.1-480513/${host_platform}/libcxx-ndk-480513-${host_platform}-x86_64.tar.bz2
+prebuilts/gcc/linux-x86/esp,${tool_repo}/openharmony/compiler/gcc_esp/2019r2-8.2.0/linux/esp-2019r2-8.2.0.zip
+prebuilts/gcc/linux-x86/csky,${tool_repo}/openharmony/compiler/gcc_csky/v3.10.29/linux/csky-v3.10.29.tar.gz
+"""
+
+darwin_copy_config="""
+prebuilts/previewer/darwin,${tool_repo}/openharmony/develop_tools/previewer/3.1.5.4/previewer-3.1.5.4.mac.tar.gz
+prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/openharmony/compiler/clang/10.0.1-480513/${host_platform}/libcxx-ndk-480513-${host_platform}-x86_64.tar.bz2
 """
 
 if [[ "${host_platform}" == "linux" ]]; then
-    copy_config+="""
-        prebuilts/cmake,${tool_repo}/harmonyos/compiler/cmake/3.16.5/windows/cmake-windows-x86-3.16.5.tar.gz
-        prebuilts/mingw-w64/ohos/linux-x86_64,${tool_repo}/harmonyos/compiler/mingw-w64/7.0.0/clang-mingw.tar.gz
-        prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi,${tool_repo}/harmonyos/compiler/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi/1.0/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi.tar.gz
-        prebuilts/gcc/linux-x86/aarch64,${tool_repo}/harmonyos/compiler/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi/1.0/gcc-linaro-7.5.0-2019.12-x86_64_aarch64-linux-gnu.tar.xz
-        prebuilts/previewer/windows,${tool_repo}/harmonyos/develop_tools/previewer/3.1.5.4/previewer-3.1.5.4.win.tar.gz
-        prebuilts/clang/ohos/windows-x86_64,${tool_repo}/harmonyos/compiler/clang/10.0.1-480513/windows/clang-480513-windows-x86_64.tar.bz2
-        prebuilts/clang/ohos/windows-x86_64,${tool_repo}/harmonyos/compiler/clang/10.0.1-480513/windows/libcxx-ndk-480513-windows-x86_64.tar.bz2
-        prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/harmonyos/compiler/clang/10.0.1-480513/${host_platform}/libcxx-ndk-480513-${host_platform}-x86_64.tar.bz2
-        prebuilts/gcc/linux-x86/esp,${tool_repo}/harmonyos/compiler/gcc_esp/2019r2-8.2.0/linux/esp-2019r2-8.2.0.zip
-        prebuilts/gcc/linux-x86/csky,${tool_repo}/harmonyos/compiler/gcc_csky/v3.10.29/linux/csky-v3.10.29.tar.gz
-        """
+    copy_config+=${linux_copy_config}
 elif [[ "${host_platform}" == "darwin" ]]; then
-    copy_config+="""
-        prebuilts/previewer/darwin,${tool_repo}/harmonyos/develop_tools/previewer/3.1.5.4/previewer-3.1.5.4.mac.tar.gz
-        prebuilts/clang/ohos/${host_platform}-x86_64,${tool_repo}/harmonyos/compiler/clang/10.0.1-480513/${host_platform}/libcxx-ndk-480513-${host_platform}-x86_64.tar.bz2
-        """
+    copy_config+=${darwin_copy_config}
 fi
 
+# download and initialize prebuild files
 if [ ! -d "${bin_dir}" ];then
     mkdir -p "${bin_dir}"
 fi
@@ -158,45 +209,49 @@ do
     if [ ! -d "${code_dir}/${unzip_dir}" ];then
         mkdir -p "${code_dir}/${unzip_dir}"
     fi
-    hwcloud_download "${bin_dir}/${md5_huaweicloud_url}.${bin_file_suffix}"  "${huaweicloud_url}"
 
-    if [ ! -f "${code_dir}/${unzip_dir}/${check_sha256}.mark" ]; then
-        if [ "X${bin_file_suffix:0-3}" = "Xzip" ];then
-            unzip -o "${bin_dir}/${md5_huaweicloud_url}.${bin_file_suffix}" -d "${code_dir}/${unzip_dir}/"
-        elif [ "X${bin_file_suffix:0-6}" = "Xtar.gz" ];then
-            tar -xvzf "${bin_dir}/${md5_huaweicloud_url}.${bin_file_suffix}"  -C  "${code_dir}/${unzip_dir}"
+    check_sha256_by_mark ${huaweicloud_url}
+    if [ ${sha256_result} -gt 0 ]; then
+        hwcloud_download "${bin_dir}/${md5_huaweicloud_url}.${bin_file}"  "${huaweicloud_url}"
+        if [ "X${bin_file:0-3}" = "Xzip" ];then
+            unzip -o "${bin_dir}/${md5_huaweicloud_url}.${bin_file}" -d "${code_dir}/${unzip_dir}/"
+        elif [ "X${bin_file:0-6}" = "Xtar.gz" ];then
+            tar -xvzf "${bin_dir}/${md5_huaweicloud_url}.${bin_file}"  -C  "${code_dir}/${unzip_dir}"
         else
-            tar -xvf "${bin_dir}/${md5_huaweicloud_url}.${bin_file_suffix}"  -C  "${code_dir}/${unzip_dir}"
+            tar -xvf "${bin_dir}/${md5_huaweicloud_url}.${bin_file}"  -C  "${code_dir}/${unzip_dir}"
+        fi
+        # it is used to handle some redundant files under prebuilts path
+        # todo: remove redundant files before prebuilts_download
+        if [ -d "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi" ];then
+            mv "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi" "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi2/"
+            rm -rf "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi"
+            mv "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi2/" "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/"
+        fi
+        if [ -d "${code_dir}/prebuilts/clang/ohos/windows-x86_64/clang-480513" ];then
+            rm -rf "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm"
+            mv "${code_dir}/prebuilts/clang/ohos/windows-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm"
+            ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm/lib/clang/current"
+        fi
+        if [ -d "${code_dir}/prebuilts/clang/ohos/linux-x86_64/clang-480513" ];then
+            rm -rf "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm"
+            mv "${code_dir}/prebuilts/clang/ohos/linux-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm"
+            ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm/lib/clang/current"
+        fi
+        if [ -d "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/clang-480513" ];then
+            rm -rf "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm"
+            mv "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm"
+            ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm/lib/clang/current"
+        fi
+        if [ -d "${code_dir}/prebuilts/gcc/linux-x86/esp/esp-2019r2-8.2.0/xtensa-esp32-elf" ];then
+            chmod 755 "${code_dir}/prebuilts/gcc/linux-x86/esp/esp-2019r2-8.2.0" -R
         fi
         echo 0 > "${code_dir}/${unzip_dir}/${check_sha256}.mark"
     fi
-
-    # it is used to handle some redundant files under prebuilts path
-    if [ -d "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi" ];then
-        mv "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/prebuilts_gcc_linux-x86_arm_gcc-linaro-7.5.0-arm-linux-gnueabi" "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi2/"
-        rm -rf "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi"
-        mv "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi2/" "${code_dir}/prebuilts/gcc/linux-x86/arm/gcc-linaro-7.5.0-arm-linux-gnueabi/"
-    fi
-    if [ -d "${code_dir}/prebuilts/clang/ohos/windows-x86_64/clang-480513" ];then
-        rm -rf "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm"
-        mv "${code_dir}/prebuilts/clang/ohos/windows-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm"
-    ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/windows-x86_64/llvm/lib/clang/current"
-    fi
-    if [ -d "${code_dir}/prebuilts/clang/ohos/linux-x86_64/clang-480513" ];then
-        rm -rf "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm"
-        mv "${code_dir}/prebuilts/clang/ohos/linux-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm"
-	ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/linux-x86_64/llvm/lib/clang/current"
-    fi
-    if [ -d "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/clang-480513" ];then
-        rm -rf "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm"
-        mv "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/clang-480513" "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm"
-	ln -snf 10.0.1 "${code_dir}/prebuilts/clang/ohos/darwin-x86_64/llvm/lib/clang/current"
-    fi
-    if [ -d "${code_dir}/prebuilts/gcc/linux-x86/esp/esp-2019r2-8.2.0/xtensa-esp32-elf" ];then
-        chmod 755 "${code_dir}/prebuilts/gcc/linux-x86/esp/esp-2019r2-8.2.0" -R
-    fi
+    echo "k"
 done
 
+
+# npm env setup and install
 
 node_js_ver=v12.18.4
 node_js_name=node-${node_js_ver}-${host_platform}-x64
@@ -208,80 +263,22 @@ if [ ! -f "${node_js_pkg}" ]; then
     tar zxf ${node_js_pkg}
 fi
 
-if [ ! -d "${code_dir}/third_party/jsframework" ]; then
-    echo "${code_dir}/third_party/jsframework not exist, it shouldn't happen, pls check..."
-else
-    cd ${code_dir}/third_party/jsframework/
-    export PATH=${code_dir}/prebuilts/build-tools/common/nodejs/${node_js_name}/bin:$PATH
-    npm config set registry ${npm_registry}
-    if [ "X${SKIP_SSL}" == "XYES" ];then
-        npm config set strict-ssl false
-    fi
-    npm cache clean -f
-    npm install
+npm_install_config="""
+third_party/jsframework,prebuilts/build-tools/common/js-framework
+developtools/ace-ets2bundle/compiler,
+developtools/ace-js2bundle/ace-loader,
+ark/ts2abc/ts2panda,prebuilts/build-tools/common/ts2abc
+"""
 
-    cd ${code_dir}
-    if [ -d "${code_dir}/prebuilts/build-tools/common/js-framework" ]; then
-        echo -e "\n"
-        echo "${code_dir}/prebuilts/build-tools/common/js-framework already exist, it will be replaced with node-${node_js_ver}"
-        /bin/rm -rf ${code_dir}/prebuilts/build-tools/common/js-framework
-        echo -e "\n"
-    fi
-
-    mkdir -p ${code_dir}/prebuilts/build-tools/common/js-framework
-    /bin/cp -R ${code_dir}/third_party/jsframework/node_modules ${code_dir}/prebuilts/build-tools/common/js-framework/
-fi
-
-if [ ! -d "${code_dir}/developtools/ace-ets2bundle/compiler" ]; then
-    echo "${code_dir}/developtools/ace-ets2bundle/compiler not exist, it shouldn't happen, pls check..."
-else
-    cd ${code_dir}/developtools/ace-ets2bundle/compiler
-    export PATH=${code_dir}/prebuilts/build-tools/common/nodejs/${node_js_name}/bin:$PATH
-    npm config set registry ${npm_registry}
-    if [ "X${SKIP_SSL}" == "XYES" ];then
-        npm config set strict-ssl false
-    fi
-    npm cache clean -f
-    npm install
-fi
-
-
-if [ ! -d "${code_dir}/developtools/ace-js2bundle/ace-loader" ]; then
-    echo "${code_dir}/developtools/ace-js2bundle/ace-loader not exist, it shouldn't happen, pls check..."
-else
-    cd ${code_dir}/developtools/ace-js2bundle/ace-loader
-    export PATH=${code_dir}/prebuilts/build-tools/common/nodejs/${node_js_name}/bin:$PATH
-    npm config set registry ${npm_registry}
-    if [ "X${SKIP_SSL}" == "XYES" ];then
-        npm config set strict-ssl false
-    fi
-    npm cache clean -f
-    npm install
-fi
-
-
-if [ -d "${code_dir}/ark/ts2abc/ts2panda" ]; then
-    cd ${code_dir}/ark/ts2abc/ts2panda
-    export PATH=${code_dir}/prebuilts/build-tools/common/nodejs/${node_js_name}/bin:$PATH
-    npm config set registry ${npm_registry}
-    if [ "X${SKIP_SSL}" == "XYES" ];then
-        npm config set strict-ssl false
-    fi
-    npm cache clean -f
-    npm install
-
-    cd ${code_dir}
-    if [ -d "${code_dir}/prebuilts/build-tools/common/ts2abc" ]; then
-        echo -e "\n"
-        echo "${code_dir}/prebuilts/build-tools/common/ts2abc already exist, it will be replaced with node-${node_js_ver}"
-        /bin/rm -rf ${code_dir}/prebuilts/build-tools/common/ts2abc
-        echo -e "\n"
-    fi
-
-    mkdir -p ${code_dir}/prebuilts/build-tools/common/ts2abc
-    /bin/cp -rf ${code_dir}/ark/ts2abc/ts2panda/node_modules ${code_dir}/prebuilts/build-tools/common/ts2abc/
-fi
-
+for i in $(echo ${npm_install_config})
+do
+    code_path=$(echo $i|awk -F ',' '{print $1}')
+    modules_path=$(echo $i|awk -F ',' '{print $2}')
+    npm_install ${code_path}
+    echo ${code_path}
+    echo ${modules_path}
+    node_modules_copy ${code_path} ${modules_path}
+done
 
 cd ${code_dir}
 echo -e "\n"
